+++ pcs_test/tier0/lib/commands/resource/test_resource_move_autoclean.py
@@ -1571,3 +1571,51 @@ class MoveAutocleanFailures(MoveAutocleanCommonSetup):
                 )
             ]
         )
+
+    def assert_moving_bundle_with_more_instances(self, stderr, report_code):
+        # This tests that error messages from crm_resources are translated to
+        # report messages correctly. For the purpose of the test, no bundle is
+        # needed. The only thing that matters is the error message from
+        # crm_resource. So we don't bother creating a CIB with a bundle in it.
+        self.tmp_file_mock_obj.set_calls(
+            self.get_tmp_files_mocks(
+                _simulation_transition_fixture(
+                    _simulation_synapses_fixture(self.resource_id)
+                )
+            )[:1]
+        )
+        self.config.runner.pcmk.load_state(
+            resources=_state_resource_fixture(
+                self.resource_id, "Started", "node1"
+            ),
+        )
+        self.config.runner.pcmk.resource_move(
+            resource=self.resource_id,
+            env=dict(CIB_file=self.cib_rsc_move_tmp_file_name),
+            returncode=1,
+            stderr=stderr,
+        )
+
+        self.env_assist.assert_raise_library_error(
+            lambda: move_autoclean(
+                self.env_assist.get_env(),
+                self.resource_id,
+            ),
+            [
+                fixture.error(report_code, resource_id=self.resource_id),
+            ],
+            expected_in_processor=False,
+        )
+        self.env_assist.assert_reports(self.get_reports(stage=1))
+
+    def test_moving_bundle_with_more_instances_1(self):
+        self.assert_moving_bundle_with_more_instances(
+            "Error performing operation: Multiple items match request",
+            reports.codes.CANNOT_MOVE_RESOURCE_MULTIPLE_INSTANCES,
+        )
+
+    def test_moving_bundle_with_more_instances_2(self):
+        self.assert_moving_bundle_with_more_instances(
+            "Resource 'A' not moved: active in 2 locations.",
+            reports.codes.CANNOT_MOVE_RESOURCE_MULTIPLE_INSTANCES_NO_NODE_SPECIFIED,
+        )